Added modular vending machine and minimal clothing support
Vending machines needed a little more code to make the russ_products and such work as intended. I simply copy pasta'd hippies code and changed references to hippie to russ. Clothing support just makes it so that things in the jumpsuit slot will get the icons from the right spot if they are made according to the modular system.